+++ src/middlewares/CurrentPageNumber.ts
@@ -10,44 +10,46 @@ import { MiddlewareFunction } from "../types/MiddlewareFunction.js";
 // Middleware
 //
 
-export type MiddlewareFritterContext = FritterContext &
+export type MiddlewareFritterContext<T> = FritterContext &
 {
 	currentPageNumber: number;
+	getPagination: () => T;
 };
 
-export type CreateOptions =
+export type CreateOptions<T = null> =
 {
-	getPageNumber?: (context: MiddlewareFritterContext) => number;
+	getPageNumber?: (context: MiddlewareFritterContext<T>) => number;
+	getPagination?: (context: MiddlewareFritterContext<T>) => T;
 };
 
-export type CreateResult =
+export type CreateResult<T> =
 {
-	getPageNumber: (context: MiddlewareFritterContext) => number;
-	execute: MiddlewareFunction<MiddlewareFritterContext>;
+	getPageNumber: (context: MiddlewareFritterContext<T>) => number;
+	getPagination: (context: MiddlewareFritterContext<T>) => T;
+	execute: MiddlewareFunction<MiddlewareFritterContext<T>>;
 };
 
-export function create(options: CreateOptions = {}): CreateResult
+export function create<T = null>(options: CreateOptions<T> = {}): CreateResult<T>
 {
-	const currentPageNumberMiddleware: CreateResult =
+	const currentPageNumberMiddleware: CreateResult<T> =
 	{
 		getPageNumber: options.getPageNumber ??
 			((context) =>
 			{
 				let currentPage = parseInt(context.fritterRequest.getSearchParams().get("page") ?? "1");
 
-				if (isNaN(currentPage))
-				{
-					currentPage = 1;
-				}
-
-				return currentPage;
+				return isNaN(currentPage) ? 1 : currentPage;
 			}),
+
+		getPagination: options.getPagination ?? (() => null as T),
 
 		execute: async (context, next) =>
 		{
 			const currentPageNumber = currentPageNumberMiddleware.getPageNumber(context);
 
 			context.currentPageNumber = currentPageNumber;
+
+			context.getPagination = () => currentPageNumberMiddleware.getPagination(context);
 
 			await next();
 		},